fix(taskHandler): build the failure bitmask with a 32-bit shift operand
The failure bitmask was built with `1U << i`. `1U` is `unsigned int`, which is
16-bit on AVR (atmega328p: __SIZEOF_INT__ == 2), so for a task index >= 16 the
shift is undefined behaviour on the AVR builds -- yet the static_assert allows up
to 32 tasks, and the bits 16..31 of the mask would be wrong. ESP (xtensa, 32-bit
int) was unaffected, and the native test could never reveal it because host int
is >= 32-bit, so test_32_tasks_highest_bit_bitmask passed while certifying a case
that would have been broken on the real 16-bit-int target.

Shift a uint32_t operand instead (static_cast<uint32_t>(1) << i). uint32_t is
>= 32-bit on every target and outranks int, so it never promotes down to a 16-bit
int -- the shift is well-defined up to bit 31 on AVR, ESP and host alike. No
current firmware trips this (alert/irrigation run 5-6 tasks), so it is a latent
contract fix, not an active bug. A comment on the 32-task test records that the
native run cannot prove the AVR behaviour; the 32-bit operand is what makes it hold.

Full release gate (build 7 envs + test 442 + per-env check) is clean.

+++ lib/taskHandler/src/taskHandler.hpp
@@ -67,7 +67,10 @@ public:
       for(uint8_t i = 0U; i < taskNum; ++i) {
         if(taskList[i] != nullptr) {
           if(!taskList[i]->init()) {
-            failureMask |= (1U << i);
+            // static_cast<uint32_t>: AVR `int` is 16-bit, so a plain `1U << i` would be undefined
+            // for i >= 16; a 32-bit operand keeps the shift well-defined up to bit 31 (taskNum <= 32)
+            // on every target. uint32_t never promotes down to int, so no narrowing on AVR either.
+            failureMask |= (static_cast<uint32_t>(1) << i);
           }
         }
       }
@@ -113,7 +116,7 @@ private:
     uint32_t failureMask = 0U;
     if(taskList[currentTask] != nullptr) {
       if(!taskList[currentTask]->run()) {
-        failureMask |= (1U << currentTask);
+        failureMask |= (static_cast<uint32_t>(1) << currentTask);  // 32-bit operand: AVR int is 16-bit (see initTasks)
       }
     }
     if(++currentTask >= taskNum) { currentTask = 0U; }
@@ -132,7 +135,7 @@ private:
     if(++currentTask >= taskNum) { currentTask = 1U; }
     if(taskList[currentTask] != nullptr) {
       if(!taskList[currentTask]->run()) {
-        failureMask |= (1U << currentTask);
+        failureMask |= (static_cast<uint32_t>(1) << currentTask);  // 32-bit operand: AVR int is 16-bit (see initTasks)
       }
     }
     return failureMask;
